Cleaned warnings. Acquisition (without saving) completed.

+++ utils.py
@@ -1,4 +1,3 @@
-import time
 from threading import Event, Timer
 import logging
 
@@ -6,7 +5,7 @@ import logging
 logger = logging.getLogger(__name__)
 
 
-def interval_polling(stop_event: Event, callback, interval_s: int, kill_event: Event):
+def interval_polling(stop_event: Event, callback, interval_s: float, kill_event: Event):
     if kill_event is not None and kill_event.is_set():
         return
     callback()
@@ -17,6 +16,6 @@ def interval_polling(stop_event: Event, callback, interval_s: int, kill_event: E
         logger.debug("Interval polling stopped!")
 
 
-def start_interval_polling(stop_event: Event, callback, interval_s: int, kill_event: Event):
+def start_interval_polling(stop_event: Event, callback, interval_s: float, kill_event: Event):
     logger.debug(f"Starting new interval polling with {interval_s}s refresh")
     Timer(interval_s, interval_polling, [stop_event, callback, interval_s, kill_event]).start()
